# api/openapi/nodal.v1.yaml
openapi: 3.1.0
info:
  title: No-dal Community Edition API
  version: 0.1.6
  description: Phase 7 lab QEMU prototype plus Phase 6 northbound API. Host visibility, Directory storage, safe networking, system containers, Terminal, and Files.
servers:
  - url: /api/v1
paths:
...
  /setup/claim:
    post:
      operationId: claimSetup
      summary: Claim the one-time setup token and create the first admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupClaimRequest"
      responses:
        "200":
          description: First administrator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
  /auth/login:
    post:
      operationId: login
      summary: Create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
...
  /tokens:
    post:
      operationId: createToken
      summary: Create an API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTokenRequest"
      responses:
        "201":
          description: Token shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateTokenResponse"
  /tokens/revoke:
    post:
      operationId: revokeToken
      summary: Revoke an API token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeTokenRequest"
      responses:
        "204":
          description: Revoked
...
  /nodes/{id}:
    get:
      operationId: getNode
      summary: One node summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Node
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeSummary"
  /nodes/{id}/hardware:
    get:
      operationId: getNodeHardware
      summary: Cached observed hardware
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Hardware inventory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HardwareResponse"
  /nodes/{id}/capabilities:
    get:
      operationId: getNodeCapabilities
      summary: Observed capability flags
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Capabilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CapabilitiesResponse"
...
  /storage/pools:
    get:
      operationId: listStoragePools
      summary: List Directory storage pools
      responses:
        "200":
          description: Pools
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoragePoolListResponse"
    post:
      operationId: createStoragePool
      summary: Create a Directory storage pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStoragePoolRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoragePool"
  /storage/pools/{id}:
    get:
      operationId: getStoragePool
      summary: Get a Directory storage pool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoragePool"
  /storage/volumes:
    get:
      operationId: listStorageVolumes
      summary: List volumes
      responses:
        "200":
          description: Volumes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageVolumeListResponse"
    post:
      operationId: createStorageVolume
      summary: Create a volume by UUID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStorageVolumeRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageVolume"
  /storage/volumes/{id}:
    get:
      operationId: getStorageVolume
      summary: Get a volume by UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Volume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageVolume"
...
  /storage/images/{id}:
    get:
      operationId: getStorageImage
      summary: Get a library item by UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageImage"
  /networks:
    get:
      operationId: listNetworks
      summary: List network objects
      responses:
        "200":
          description: Networks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkListResponse"
    post:
      operationId: createNetwork
      summary: Create or dry-run a network
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNetworkRequest"
      responses:
        "200":
          description: Dry-run preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NetworkPreview"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "409":
          description: Typed confirmation required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConfirmationRequired"
  /networks/{id}:
    get:
      operationId: getNetwork
      summary: Get a network by UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
  /networks/{id}/apply:
    post:
      operationId: applyNetwork
      summary: Apply or dry-run a network
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: dry_run
          in: query
          schema:
            type: boolean
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyNetworkRequest"
      responses:
        "200":
          description: Applied or previewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
  /networks/{id}/reservations:
    get:
      operationId: listNetworkReservations
      summary: List DHCP reservations
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReservationListResponse"
    post:
      operationId: createNetworkReservation
      summary: Create a DHCP reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReservationRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /workloads:
    get:
      operationId: listWorkloads
      summary: List system containers
      responses:
        "200":
          description: Workloads
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkloadListResponse"
    post:
      operationId: createWorkload
      summary: Create a system container
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkloadRequest"
      responses:
        "200":
          description: Idempotent replay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
  /workloads/{id}:
    get:
      operationId: getWorkload
      summary: Get a workload by UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
    patch:
      operationId: updateWorkload
      summary: Edit cpu, memory, or desired power
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkloadRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
    post:
      operationId: updateWorkloadPost
      summary: Edit cpu, memory, or desired power
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkloadRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
  /workloads/{id}/start:
    post:
      operationId: startWorkload
      summary: Start a system container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
  /workloads/{id}/stop:
    post:
      operationId: stopWorkload
      summary: Stop a system container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
  /workloads/{id}/restart:
    post:
      operationId: restartWorkload
      summary: Restart a system container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
  /workloads/{id}/delete:
    post:
      operationId: deleteWorkload
      summary: Delete a system container runtime
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
  /workloads/{id}/clone:
    post:
      operationId: cloneWorkload
      summary: Clone a system container
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloneWorkloadRequest"
      responses:
        "201":
          description: Cloned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workload"
  /nodes/{id}/terminal/sessions:
    post:
      operationId: createNodeTerminalSession
      summary: Create an admin-only host terminal ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIOSessionRequest"
      responses:
        "201":
          description: Ticket shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IOSession"
  /workloads/{id}/terminal/sessions:
    post:
      operationId: createWorkloadTerminalSession
      summary: Create a CT terminal ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIOSessionRequest"
      responses:
        "201":
          description: Ticket shown once
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IOSession"
...
components:
  schemas:
...
    SetupClaimRequest:
      type: object
      additionalProperties: false
      required:
        - token
        - username
        - password
      properties:
        token:
          type: string
        username:
          type: string
        password:
          type: string
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
    MeResponse:
      type: object
      additionalProperties: false
      required:
        - user_id
        - username
        - roles
        - edition
      properties:
        user_id:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        edition:
          type: string
        cluster_id:
          type: string
    CreateTokenRequest:
      type: object
      additionalProperties: false
      required:
        - name
      properties:
        name:
          type: string
    CreateTokenResponse:
      type: object
      additionalProperties: false
      required:
        - id
        - prefix
        - token
      properties:
        id:
          type: string
        prefix:
          type: string
        token:
          type: string
    RevokeTokenRequest:
      type: object
      additionalProperties: false
      required:
        - id
      properties:
        id:
          type: string
...